+++ x/evm/types/testutils/types.go
@@ -326,7 +326,7 @@ func RandomParams() types.Params {
 		MinVoterCount:       rand.PosI64(),
 		CommandsGasLimit:    uint32(rand.I64Between(0, 10000000)),
 		EndBlockerLimit:     rand.PosI64(),
-		TransferLimit:       rand.PosI64(),
+		TransferLimit:       uint64(rand.PosI64()),
 	}
 
 	params.Network = params.Networks[int(rand.I64Between(0, int64(len(params.Networks))))].Name
